# openapi.yml
openapi: 3.0.3
info:
  title: Viable Stone Paint Shop API
  version: 1.0.0
  description: Internal management API for the Viable Stone paint and painting-equipment
    shop. Source of truth for authentication, prices, stock, sales, payments, receipts,
    expenses, approvals and reports.
paths:
...
  /api/v1/inventory/low-stock/:
    get:
      operationId: inventory_low_stock_retrieve
      description: Branch balances at or below each variant's low-stock level. Paginated.
        Cashiers get quantity + price; an owner also gets cost fields at runtime (the
        schema documents the guaranteed subset).
      summary: Low-stock balances (paginated)
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - Inventory
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventoryBalanceEmployeeList'
          description: ''
        default:
          $ref: '#/components/responses/Error'
  /api/v1/inventory/movements/:
    get:
      operationId: inventory_movements_retrieve
      description: 'Append-only stock-movement ledger for one variant, newest first,
        paginated. Owner only; `?variant=<id>` is required. Each row: movement type,
        quantity delta, cost snapshot, actor, reference type/id and timestamp.'
      summary: Stock-movement ledger for one variant (owner, paginated)
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: variant
        schema:
          type: string
        description: Variant id whose stock-movement ledger to return.
        required: true
      tags:
      - Inventory
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStockMovementList'
          description: ''
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/inventory/stock-value/:
    get:
      operationId: inventory_stock_value_retrieve
      description: Branch stock. Cashiers see quantity + price; owner also sees cost.
      summary: Total stock value at cost (owner)
      tags:
      - Inventory
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockValue'
          description: ''
        default:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
...
    InventoryBalanceEmployee:
      type: object
      description: 'Cashier view: quantity + selling price, never cost.'
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        variant:
          type: string
          format: uuid
          readOnly: true
        variant_sku:
          type: string
          readOnly: true
        product_name:
          type: string
          readOnly: true
        variant_description:
          type: string
          readOnly: true
        quantity:
          type: integer
          readOnly: true
        low_stock_level:
          type: integer
          readOnly: true
        is_low:
          type: boolean
          readOnly: true
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,2})?$
          nullable: true
          readOnly: true
        last_restocked_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - id
      - is_low
      - last_restocked_at
      - low_stock_level
      - price
      - product_name
      - quantity
      - variant
      - variant_description
      - variant_sku
...
    MovementTypeEnum:
      enum:
      - OPENING
      - RESTOCK
      - SALE
      - RETURN
      - DAMAGE
      - ADJUSTMENT
      - OFFLINE_RECONCILIATION
      type: string
      description: |-
        * `OPENING` - Opening stock
        * `RESTOCK` - Restock
        * `SALE` - Sale
        * `RETURN` - Return
        * `DAMAGE` - Damage / write-off
        * `ADJUSTMENT` - Correction
        * `OFFLINE_RECONCILIATION` - Offline reconciliation
...
    PaginatedInventoryBalanceEmployeeList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 5
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/InventoryBalanceEmployee'
...
    PaginatedStockMovementList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 5
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/StockMovement'
...
    StockMovement:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        variant:
          type: string
          format: uuid
          readOnly: true
        variant_sku:
          type: string
          readOnly: true
        movement_type:
          allOf:
          - $ref: '#/components/schemas/MovementTypeEnum'
          readOnly: true
        quantity_delta:
          type: integer
          readOnly: true
        unit_cost_snapshot:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        reference_type:
          type: string
          readOnly: true
        reference_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        reason:
          type: string
          readOnly: true
        created_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - created_by
      - id
      - movement_type
      - quantity_delta
      - reason
      - reference_id
      - reference_type
      - unit_cost_snapshot
      - variant
      - variant_sku
    StockValue:
      type: object
      description: G5 — the ``/inventory/stock-value/`` summary (owner). Cost total
        only.
      properties:
        stock_value:
          type: string
          readOnly: true
      required:
      - stock_value
...
    Error:
      type: object
      description: Standard error envelope. Every non-2xx response has this shape.
        `code` is a stable machine-readable string that frontend logic may branch
        on; `message` is human-readable and may change; `field_errors` maps a field
        name to its messages on a 400 validation error (empty otherwise); `request_id`
        echoes `X-Request-ID` for support. Internal exception names, stack traces
        and secrets are never included.
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
        field_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        request_id:
          type: string
      required:
      - code
      - message
      - field_errors
      - request_id
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vs_sessionid
  responses:
    Error:
      description: Error envelope — 400/401/403/404/405/409/413/415/422/429/500 all
        use this shape. See the `Error` schema and FRONTEND_HANDOFF.md for the stable
        `code` catalogue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
servers:
- url: /
  description: Current host